+++ .storybook/parts/wp-styles.js
@@ -1,0 +1,6 @@
+// require('EightshiftBlocksStorybookWp/editor/build-style/editor-styles.css');
+require('EightshiftBlocksStorybookWp/editor/build-style/style-rtl.css');
+require('EightshiftBlocksStorybookWp/editor/build-style/style.css');
+require('EightshiftBlocksStorybookWp/components/build-style/style.css');
+require('EightshiftBlocksStorybookWp/block-editor/build-style/style.css');
+require('EightshiftBlocksStorybookWp/format-library/build-style/style.css');