+++ blocks/init/src/blocks/components/heading/components/heading-toolbar.js
@@ -1,7 +1,7 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
-import { AlignmentToolbar } from '@wordpress/editor';
+import { AlignmentToolbar } from '@wordpress/block-editor';
 import { Fragment } from '@wordpress/element';
-import { HeadingLevel } from 'EighshiftComponentHeadingLevel';
+import { HeadingLevel } from 'EightshiftComponentHeadingLevel';
 
 export const HeadingToolbar = (props) => {
   const {