+++ blocks/init/src/blocks/components/link/components/link-options.js
@@ -1,8 +1,8 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
 import { __ } from '@wordpress/i18n';
 import { Fragment } from '@wordpress/element';
-import { URLInput } from '@wordpress/editor';
-import { ColorPaletteCustom } from 'EighshiftComponentColorPalette';
+import { URLInput } from '@wordpress/block-editor';
+import { ColorPaletteCustom } from 'EightshiftComponentColorPalette';
 import { ToggleControl, Icon, BaseControl } from '@wordpress/components';
 import globalSettings from './../../../manifest.json';