+++ blocks/init/src/blocks/components/paragraph/components/paragraph-options.js
@@ -1,7 +1,7 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
 import { Fragment } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
-import { ColorPaletteCustom } from 'EighshiftComponentColorPalette';
+import { ColorPaletteCustom } from 'EightshiftComponentColorPalette';
 import { SelectControl, Icon } from '@wordpress/components';
 import globalSettings from './../../../manifest.json';