+++ blocks/init/src/blocks/components/paragraph/components/paragraph-toolbar.js
@@ -1,5 +1,5 @@
 import React from 'react'; // eslint-disable-line no-unused-vars
-import { AlignmentToolbar } from '@wordpress/editor';
+import { AlignmentToolbar } from '@wordpress/block-editor';
 import { Fragment } from '@wordpress/element';
 
 export const ParagraphToolbar = (props) => {