+++ blocks/init/src/blocks/custom/example/docs/story.js
@@ -1,4 +1,4 @@
-import { Gutenberg, blockDetails, hasWrapperDecorator } from 'EighshiftBlocksStorybookHelpers';
+import { Gutenberg, blockDetails, hasWrapperDecorator } from 'EightshiftBlocksStorybookHelpers';
 import React from 'react'; // eslint-disable-line no-unused-vars
 import manifest from './../manifest.json';
 import readme from './readme.md';